requirements.txt
about-time==4.2.1
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alive-progress==3.1.4
anyio==3.7.0
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
aws-sam-translator==1.64.0
boto3==1.26.111
botocore==1.29.111
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
cfn-lint==0.77.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
exceptiongroup==1.1.1
filelock==3.12.0
flake8==6.0.0
flatbuffers==23.5.26
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grapheme==0.6.0
grpcio==1.54.2
h11==0.14.0
h5py==3.8.0
httpcore==0.17.2
huggingface-hub==0.15.1
idna==3.4
imageio==2.30.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
jax==0.4.11
Jinja2==3.1.2
jmespath==1.0.1
jschema-to-python==1.2.3
jsonpatch==1.32
jsonpickle==3.0.1
jsonpointer==2.3
jsonschema==4.17.3
junit-xml==1.9
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
linkify-it-py==2.0.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mccabe==0.7.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
ml-dtypes==0.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
networkx==3.1
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.1
pbr==5.11.1
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pyflakes==3.0.1
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.1
rsa==4.9
s3transfer==0.6.0
sarif-om==1.0.4
scipy==1.10.1
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sympy==1.12
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcharts==1.1.2
termcolor==2.3.0
textual==0.27.0
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
tqdm==4.65.0
transformers==4.29.2
trio==0.22.0
trio-websocket==0.10.2
typing-extensions==4.6.3
uc-micro-py==1.0.2
urllib3==1.26.16
Werkzeug==2.3.4
wrapt==1.14.1
wsproto==1.2.0
yapf==0.33.0
yarl==1.9.2
zipp==3.15.0
